Revert "start to deprecate child class"
This reverts commit 3845fb421b1c2494ec42478bfb41715c2f7e18ea.
